[#37] Move scenario to nettest
Problem: We want to run nettest scenarios in both pure and real chain
implementation. However, currently nettest scenario is located in the
main test-suite, which cannot be easily reused for running nettest
scenario on real chain.

Solution: Move nettest scenario to `nettest` and test helpers to
`test-common` directories. `stablecoin-nettest` now runs both pure
and real chain nettest scenarios.
